feat: allow user sign in through api
This integrates the SignIn service with Tipid's GraphQL API.

This adds a lot of stuff for working with the API as well as
some other stuff that makes it possible to do automated tests
even with the API stuff added in.

Yes, this is a very large commit. Bear with me because I can't
think of any way to divide this into parts without the whole set
of changes losing sense.