ci: switch to circleci and codecov
This changes our CI provider to CircleCI and our tool for
code coverage reporting to CodeCov.

We are switching to CircleCI because Travis is very limited.
For example, we cannot use a Docker image from DockerHub as
our environment. This is an important option which is why
we are going with CircleCI instead.

We are switching to CodeCov because Coveralls has not been
stable as of late. It keeps going down for maintenance or
has an outage every now and then. CodeCov is more stable so
we are going with that instead.