refactor: improve architecture of api clients
This improves the way we are creating and using our API clients.

This removes TipidApiProvider and instead opting to use Provider
for injecting the dependency across the app.

I actually did not know that you can use Provider for things outside
of state management so this is just applying that new-found knowledge.